refactor(providers): rename contract package to providertest
Follow Go conventions for test helper packages (like net/http/httptest).
The new name makes it immediately obvious this is a test utility.
